--- requirements.txt
@@ -1,10 +1,0 @@
-certifi==2019.3.9
-joblib==0.13.2
-numpy==1.16.2
-pandas==0.24.2
-python-dateutil==2.8.0
-pytz==2018.9
-scikit-learn==0.20.3
-scipy==1.2.1
-six==1.12.0
-wincertstore==0.2
